requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.0
bleach==3.1.5
blurhash==1.1.4
boost==0.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cmake==3.17.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==2.2.12
dlib==19.20.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
Flask-MQTT==1.0.7
gast==0.3.3
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
imutils==0.5.3
incremental==17.5.0
ipykernel==5.3.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
Mastodon.py==1.5.1
matplotlib==3.0.3
mistune==0.8.4
mysql-connector-python==8.0.20
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.2.1
packaging==20.4
paho-mqtt==1.5.0
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.6.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2018.11.26
prometheus-client==0.8.0
prompt-toolkit==2.0.10
Protego==0.1.16
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.6.1
PyHamcrest==2.0.2
pymodbus==2.3.0
pyModbusTCP==0.1.8
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-magic==0.4.18
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
Scrapy==2.1.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
Twisted==20.3.0
urllib3==1.25.8
virtualenv==20.0.20
virtualenv-clone==0.5.4
w3lib==1.22.0
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==1.2.0
zope.interface==5.1.0
